🗑 Disable auto-lightbox on amp4email
